rename(square): 모델 패키지 생성 - data layer의 model 패키지 분리 (friend, question)
